requirements.txt
Django>=5.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3.1
django-filter>=24.1
mysqlclient>=2.2.0
drf-spectacular[sidecar]>=0.28.0